Move local state write correctness into runtime context
