+++ loopx/todos.py
@@ -12,7 +12,7 @@ from .agent_registry import (
 )
 from .file_lock import exclusive_file_lock
 from .history import load_registry
-from .local_state_write_correctness import build_todo_write_correctness_dry_run_packet
+from .control_plane.runtime.local_state_write_correctness import build_todo_write_correctness_dry_run_packet
 from .state_refresh import now_local, resolve_goal_state
 from .control_plane.goals.active_state_metadata import todo_role_for_heading
 from .status import (
